Move shape.py contents to shape_utils.py. Fix tests.
